Fix all MSVC warnings: C4114, C4003, C4005, C4996; update FreeRTOS/queue stubs and suppress fopen warnings

+++ main/dmx_preset.hpp
@@ -14,7 +14,7 @@ class DmxPreset
     DmxPreset();
 
     // Set preset name
-    void setIndex(uint16_t index) { index_ = index; }
+    void setIndex(uint16_t index) { index_ = static_cast<uint8_t>(index); }
     uint16_t getIndex() const { return index_; }
 
     void setName(const char *name);
